Fiji launcher: when falling back to system Java, unset the JAVA_HOME
If Fiji did not start because libjvm could not be loaded, the reason
might be that JAVA_HOME points to an invalid location.

When that happens, the launcher falls back to starting Fiji using the
system Java, which should know it's JAVA_HOME, so let's not interfer
with that.

Noticed and diagnosed by Jean-Marc Lugrin.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>