Coloc 2: Make MaskCursor able to repeat lower dimensional masks
This is allows one to use irregular masks without specifying a "mask stack"
for stacks. One can use a 2D mask (like drawing a circular ROI in ImageJ/
Fiji) or a 2D image for using it as mask on a stack. This saves some
memory and makes mask creation easier.