Coloc 2: Add two more irregular ROI test cases, similar to regular ones
A pixel count test has been added. For that a rectangular 2D ROI mask is
created and used as an irregular ROI on a stack. For the volume of the
"mask box" can be calculated and is asserted to be the same as if one counts
the cursor positions. The dimension properties of a MaskedImage are tested
in the same way as for regular ROIs.