Rework the updater tests
The previous tests may have been well-intentioned, but when one of them
failed, it was very hard to see what needs fixing.

So rewrite the whole bruhaha, not caring so much about speeeed, but more
on usefulness with respect to failing tests.

The strategy is this: one script (tests/gui-updater.bsh) runs all the
tests. Each test starts by an initialization step followed by an upload
step, a download step, or both. The upload/download steps start up their
own little Fijis in their own little (temporary) Fiji.app/ directories.

While the current state does not test for everything the old test did,
it is clearer to read and extend, and therefore more likely to be run
at all.

Oh, and use Beanshell to accelerate the startup.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>