Move the submodules into modules/
The move of the submodules themselves was performed by the Fiji Build
after calling

	tomove="$(git ls-files --stage | grep ^160 | grep -v java/)"

	echo "$tomove" | sed 's/^.\{50\}/&modules\//' |
	git update-index --add --index-info

	git rm --cache $(echo "$tomove" | sed 's/^.\{50\}//')

This was necessary since Git does not support submodules quite well:
It should have been as easy as

	git mv $(git submodule names | grep -v java/) modules/

(including the actual move of the directories).

Of course, also .gitmodules, the Fakefiles and build-jython.sh had
to be adjusted to account for the fact that there is now one more
level between the submodules and the Fiji launcher.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>