Add a base class for Command Launchers (and a subclass for JFrames)
The functionality in ImageJ's Command Launcher (Plugins>Utilities>Find
Commands) is just too useful to keep only for ImageJ.

Unfortunately, ImageJ's version of it is too specialized and hard to
subclass.

Fortunately, it is Open Source, so here is a ripped-up version that
provides an abstract base class for all kinds of Command Launchers,
and a specialization dedicated to JFrames which lists all entries
of all menu items.

Requested by Benjamin Schmid.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>