Coloc 2: Make MetaCursor more robust
The definitions of fwd() have been removed from MetaCursor. The implementation
was not very robust as it assumed that all cursors move in the same way over
their data. This is not true in general and therefore the ConstraintCursor
now forwards one cursor and sets the location information of the other.
This way it is made sure that both cursors refer to the same locations.

As a consequence the "repeating" of e.g. a 2D mask in a 3D stack has been
reworked. The RoiShiftingLocalizableByCursor now supports optional ignoring
of dimensions larger than the underlying images. This is made use of by the
MaskedImage to offer auto-"repeating" of a lower dimension mask.

NamedImageContainer has been made more general.