Fiji launcher (win32): work around Java3D fullscreen bug
Setting the sun.direct2d.noddraw property to true prevents -- at least
in one case -- the fullscreen mode of the 3D Viewer to be hidden by a
gray-white foreground on Windows. Let's leave this work-around in place
until we know how to fix it better.

The issue was noticed and reported by Daniel James White.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>