TrackMate_: move from JGraph to JGraphX
JGraphX is the most recent version of JGraph, a tool for the
visualization of graphs. We used JGraph here so far, because of the
bridge with JGraphT, which we use as a model for our tracks.
JGraphX is lighter, sexier and less messy; let's use it.
Going from one to another is going to be long and painfull however, this
commit is the first of a long series.
Signed-off-by: Jean-Yves Tinevez <jean-yves.tinevez@pasteur.fr>