TrackMate_: JGraph -> JGraphX: fixed vertex rendering
We moved from having many customized heavy sub-class, and preferred to
re-use the JGraphX API as muchas we could. It turned to be good enough.
Now spots are rendered with an icon on the left of the cell, and that
scales when the user changes the cell size. Good for inspection.
The 2 capture actions were buggy with our previous custom
implementation. Now they work well.
Also, the performane of the graph display is very good, thanks to the
caching in JGraphX API.
Signed-off-by: Jean-Yves Tinevez <jean-yves.tinevez@pasteur.fr>