TrackMate_: JGraph -> JGraphX: trashed custom JGraphX cells
Before this commit, we were using a specialized cell for representation
if vertices (spots) and edges (links) in JGraphX, that were extending
mxCell, and storing a reference to the Spot object they were
representing.
They were generated through a convoluted factory with its own hierarchy.

We just thrashed them all.
We now use a plain mxCell, and retrieve the corresponding spot or edge
or vice-versa through 4 maps set in the JGraphXAdapter.
Simpler and lighter.

Signed-off-by: Jean-Yves Tinevez <jean-yves.tinevez@pasteur.fr>