Unbreak ./fiji --jython on Linux
As running with LD_DEBUG=all shows, JFFI (which replaced JNA in
Jython 2.5.2) expects to find the pthread keys linked to the executable
(which we tried not to require so far).

Unfortunatley, JFFI does not fail gracefully, but with a segmentation
fault. So let's link to libpthread even if we do not use it ourselves.
At least until there is a proper fix.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>