Simple_Neurite_Tracer: initial support for tracing in the 3D Viewer
This adds a behavior to the 3D viewer, so that if you have the
wand tool selected, you can click to start a path in the 3D viewer.
(You need to switch back to the hand tool if you want to rotate
the volume again.)  Clicking in the 3D viewer with the wand tool
selected should now be equivalent to clicking in the 2.5D interface.

This support isn't complete yet, though - I still need to:

  - add a ball to mark where the selected start point is

  - (hopefully) find a way to not have to switch tool

  - understand how to make this work in noisy images with
    lots of background - often some noise in front of your
    target is selected by mistake