TrackMate_: Changed the way spot icons are generated
Now the SpotIconGrabber is a FeatureAnalyzer, though it is not linked to
any Feature (edgy here). Spot images are computed with the
FeatureFacade, as every feature is. It is stored inside a Spot object as
a base64 string, which allow its use in the track scheme straight out of
the box. We might as well consider saving spot images in the xml file in
this way.
This comes with a good performance increase in the track scheme, since
we do not recompute spots icon all the time when we do a layout.
Also, simplfy styles for mxCells in layout.
Signed-off-by: Jean-Yves Tinevez <jean-yves.tinevez@pasteur.fr>