3D Viewer: really hide consumed events from ImageJ
When an object was selected, a consumed mouse event was still passed on
to the ImageCanvas instance. For example, the Simple Neurite Tracer
wants to override the handling of mouse clicks when the want tool is
active.

We prevent this from happening by adding a behavior which relays only
unconsumed events to the listeners.

Noticed by Mark Longair, co-authored by Benjamin Schmid.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>