Simple_Neurite_Tracer: avoid focus / key shortcut confusion with Sholl analysis
There have been problems on Mac OS and other systems without
point-to-focus with people trying to start the Sholl analysis
interface while the focus isn't on the image window.  This commit
adds a key listener to detect such attempts and suggest to people
what might be wrong if they type Shift with Control-A or Alt-A
in the wrong window.