TrackMate_: put back matrix pruning optimization
Fix a bug in and turn back on Nick's optimization based on cost matrix
pruning. The idea is that rows and columns containing only blocking
values are useless, since they will not yield any link anyway.
So they should be removed, generating a smaller cost matrix, which in
turns yields a much faster tracking process (since it depends in O(n^3)
or O(n^4) where n is the matrix size).

As expected, there is a marked improvement for large dataset, such as 2h
of C.elegans embryo lineage (in O(n^4)):

- Non optimized case:
Cost matrix for frame-to-frame linking created in 0.0 s.
Frame to frame LAP solved in 0.2 s.
Cost matrix for track segments created in 0.4 s.
Track segment LAP solved in 12.0 s.

- Optimized case:
Cost matrix for frame-to-frame linking created in 0.0 s.
Frame to frame LAP solved in 0.2 s.
Cost matrix for track segments created in 0.1 s.
Track segment LAP solved in 0.1 s.

Signed-off-by: Jean-Yves Tinevez <jean-yves.tinevez@pasteur.fr>