TrackMate_: dsign change: separate model and plugin
Now, the plugin has a model.
The plugin contains all the methods to process the model (all the
execSomething() methods), whereas the model has just the methods to
retrieve data and eventually modify it.
The views uses the model.
The controller pilots a plugin. So the GUI is just a clickodrome for the
plugin.
With this, one can still make a full tracking procedure in command line
with a single class (the plugin), but this frees a bit the model, which
had all the code in it.
Signed-off-by: Jean-Yves Tinevez <jean-yves.tinevez@pasteur.fr>