TrackMate_: properly handle track visibility when linking manually
Now that we have tracks that can be visible or not, we have an extra
layer of complexity when doing manual editing. The root of this
complexity is lying in the track scheme.

After tracking, we can have a very large number of tracks, summing to a
even larger number of spots in these tracks. Fortunately, we can filter
these tracks, and only retain those which are valuable, hopefully in a
smaller number.

TrackScheme can be slow when it receives a large number of tracks and
spots, because it needs to create at least one cell per spot/edge. So at
creation time, we do not feed it the whole track collection, but only
those retained after filtering. Cells are created only for these tracks, and
other spots and edges are unknown to TrackScheme.

But now what happens if the user doing manual editing in TrachScheme
decides to link a spot from a visible track with a spot from an
invisible track? Or if he decides to create a link between two spots,
when this link actually already exists and lies in an invisible track?

This causes a whole mess, hopefully sorted out by this commit. The
following rules were implemented:

- If a link is made between at least one spot which belong to an
  invisible track, the whole track is made visible and imported into
  TrackScheme.
- If a link is made and that link already exists, but was invisible, it
  is retrieved and imported in TrachScheme, and the whole track is
  made visible and imported into TrackScheme.
- If a spot is created, it is imported as standalone in TrackScheme.
- If a spot is modified, is is imported in TrackScheme, but not the
  track it belongs to (standalone).

That can generate surprising results to the user, if we do not warn him
wih a proper documentation, that indeed needs to be written. He. A bit
like the apidoc of the TrackMateModel, which became quite complicated
(for me) since it is in charge of handling these events.

Note that this complexity arose because we want to have something which
is convenient to manually edit.

Signed-off-by: Jean-Yves Tinevez <jeanyves.tinevez@gmail.com>