TrackMate_: fix a nasty bug: spots with duplicate IDs
Now that we multi-threaded everything we could, we of course run into
concurrent problems. With the naive Spot ID counter we had, based on
int, we could have spots with duplicate IDs, which would fsck the tracks
when loading/saving a file, generating all sort of bad edges.

Fix this using an AtomicInteger as a counter. Test attached to this
commit revealt the problem and its solution.
Signed-off-by: Jean-Yves Tinevez <jean-yves.tinevez@pasteur.fr>