Script Editor: add a convenience plugin to call Fiji Build
In the Fiji Updater, we want to be able to rebuild a plugin when "Show
changes" said that it contains debug information or untracked files. So
let's add that functionality (as with "Show changes", the worker class
lives in the Script Editor since we want to avoid depdendencies for the
Updater and developers have the Editor anyway).

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>