3D Viewer (STL import): try harder to detect whether the file is binary
Some jokers start the comment in the binary STL with "solid <name>" just
as if they would write an ASCII STL file. SolidWorks does that, too.

However, a pretty good rule is that a NUL character is not really ASCII,
so let's test for any NUL in the first "line" of the file, too.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>