TrackMate_: fix a bug in 3D displayer
At instantiation, the 3D displayer showed all spots visible, ignoring
the filteredSpots fields, which sets the visibility of individual
spots.
This is fixed by this commit.

Also we fix a potential bug in the GUI process: if the displayer
rendering step was taking too long (compared to calculating the spots
features, which happens simultaneously), the user could press 'next'
while the displayer was not fully rendered, generating a NPE in the next
step.
Signed-off-by: Jean-Yves Tinevez <jean-yves.tinevez@pasteur.fr>