TrackMate_: decided that a restriction was not a bug
The current trackers do not allow to detect track starting by a split
or ending in a merge: they allow for Ys, but not for Vs.
Also, they do not detect complex points: no Xs, but you can have IIs.
I decided that it was actually acceptable from a concrete viewpoint.
This needs to be stated in the doc however.
Signed-off-by: Jean-Yves Tinevez <jeanyves.tinevez@gmail.com>