TrackMate_: fix a bug with the SpotEditTool
If the TrackMate plugin was launched several time in one session, or
even was used to load a file while the SpotEditTool was active, new
image windows were not recognized by the tool.

It seems like they were not registering the tool as a listener for them.
This commit fixes that.
Signed-off-by: Jean-Yves Tinevez <jeanyves.tinevez@gmail.com>