Change the license of the Script Editor to BSD
This only refers to the source code in Fiji, of course, and does not
pretend to extend to the RSyntaxTextArea which is licensed under the LGPL.
In the near future, the Script Editor will be refactored so it can run
with barebones JTextArea optionally.

Here are excerpts of the mails with the consent of the other authors of
the Script Editor:

From: Albert Cardona <EMAIL ADDRESS>
Date: Wed, 07 Sep 2011 20:42:49 +0200

Hi Johannes,

changing the license of parts or the whole Script Editor to BSD is fine
with me. The Script Editor is a core component of Fiji and should
definitely make it into ImageJ 2.0.

Albert

Date: Thu, 8 Sep 2011 10:01:08 +0200
From: Tom Kazimiers <EMAIL ADDRESS>

Hi Johannes,

On Wed, Sep 07, 2011 at 10:36:23AM -0500, Johannes Schindelin wrote:
> For visibility reasons, I plan to make a stripped-down version that will
> make it into ImageJ2. For this, I would like to ask you for your okay to
> relicense under the BSD license.

the BSD relicensing is okay for me. Great, that the script editor will
make it into ImageJ2.

Date: Thu, 8 Sep 2011 15:58:01 +0530
From: Sumit Dubey <EMAIL ADDRESS>

Hi Johannes,

Yes, the change is okay with me. Nice hearing from you again.

Thanks,
Sumit

From: Jacques_Pécréaux <EMAIL ADDRESS>
Date: Thu, 8 Sep 2011 09:40:52 +0200

Dear Johannes,

Script editor is a strong advantage of fiji over ImageJ 1. So, it would
be great to put it in ImageJ 2. And thank you for taking care of that!
My contribution was anyway minute....

[...]

Jacques

From: Jean-Yves Tinevez <EMAIL ADDRESS>
Date: Wed, 7 Sep 2011 18:06:09 +0200

Dear Johannes,

You have my agreement for the BSD license for the Fiji Script Editor.
[...]

Best regards, and congrats on your new job! Yeah!
jy

Date: Thu, 8 Sep 2011 13:10:39 +0800
From: Yili Zhao <EMAIL ADDRESS>

Hi,
  I'm ok for the change, and thank you for the effort mading ImageJ2 better :)

[...]

--
Yili Zhao

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>