Javascript: support obsolete Rhino script engines
Sometimes, you just cannot help things. For example, when a completely
outdated Javascript engine sits in the non-removable part of the
classpath.

In the good Fiji spirit, let's not accept this to break our support for
Javascript and bend over by doing some low-level magic to accomodate
Javascript engines that do not understand setApplicationClassLoader() and
therefore might well try a class loader that has not the faintest clue
where Fiji's classes are located.

Noticed and fixed while working on a bug report by Walter de Back.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>