Add a convenience class: fiji.Debug
The only purpose of this class is to make debugging easier by delegating
to the Object_Inspector.

For the moment, it contains two methods, one to show the object
inspector's frame on an arbitrary object, and the other to access
non-public fields, both via reflection.

For example, if you need to print the current contents of, say, the img
field of ImagePlus, you can write:

	IJ.log("img: " + fiji.Debug.get(imp, "img"));

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>