Coloc 2: fix reliance on IJ1 side effect
When showing a new ImageWindow in IJ1, a substantial amount of time goes
by until the image is actually drawn. Switching to the JImagePanel
uncovered a bug where Coloc 2 relied on that repaint coming way after
toggling logarithmic display of the mask.

Hunting this bug down took this developer a complete workday since it was
not clear whether it is a bug in IJ1 or in Coloc 2, together with timing
issues regarding simultanous threads.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>