TrackMate_: WIP: de-enumify segmenters
We continue our crusade against enum types for concrete implementations.
Our holy land will be an easility extendable plugin.
Today, our march brings us along the coasts of segmenters. They are
nastily worshiping enums, and we are here to bring them back to the holy
light of coding to interface.
Signed-off-by: Jean-Yves Tinevez <jean-yves.tinevez@pasteur.fr>