TrackMate_: WIP: de-enumification of segmenters
I am right now in the tricky part of GUI interaction, which we still
want to automate. Another difficulty is with IO: it turns out that a
segmenter settings object must be able to generate its own XML code so
as to be saved, in a portable way.

So we will have not-so good looking code, where a segmenter settings
will have some code to generate a Swing GUI that can configure it, and a
JDom related code to generate the XML that can save it...
Signed-off-by: Jean-Yves Tinevez <jean-yves.tinevez@pasteur.fr>