Start a new class, IJHacker, to eventually cope with plain IJ1
Using Javassist, we will be able to use plain ImageJ rather than
roll our ever diverging ImageJA forever. This class is just a start;
when it is done, we will overwrite ImageJA's current state with
a plain ImageJ1's one and ship with that.

Of course, another advantage is that everybody can update to ImageJ's
daily build then without running into problems.

For now, the IJHacker only adds special handling to warn specifically
about dependencies that were not found.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>