IJHacker: call Fiji's Script Editor instead of IJ1's TextEditor
The Script Editor is so much nicer, and we still want to have all the
goodness even if we will switch to plain ImageJ soon.

Okay, this is it. This is how the class merits its name. It is a total
hack, especially the part where the IJ1 code constructs an Editor instance
long before it is needed and simply returns if none could be constructed.
Worse, constructing the Editor will show it right away.

And since Javassist does not really have high-level tools for editing
returns or comparisons, we have to go real low-level. The bytecode kind of
low-level.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>