Coloc 2: add a test that checks for the correctness of a masks content
For making a content test a mask is created and its content is written to
a separate image. The content of the image is then compared with the
content of the original image at the positions suggested by the mask. Both
contents should be the same.

Furthermore a small bug in a test of the predicate cursor has been fixed, a
wrong assert() method has been used.