Add a Javassist-based method to write out headless.jar
This takes plain ij.jar (ImageJA or ImageJ, does not matter) and writes
out a headless.jar.

The idea is to have this done at run-time, optionally at compile time for
setups where run-time class rewriting is not an option.

And then we actually have all Fiji-specific changes we had in ImageJA in
Fiji and can track ImageJ in the 'master' branch of ImageJA, but as a
Maven project (by moving resources and source files into the appropriate
location and adding/updating a pom.xml file).

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>