Fiji: add a base class for Javassist-backed instrumentation
This class condenses the (not IJ-specific) essence of the IJHacker
class into a base class, and adds nice helper methods for verification,
using the JustIce component of the ASM library included in our JRuby
support, and for writing out the modified classes into an own .jar
file.

This will be used for the upcoming Javassist-based headless support.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>