Switch Bio-Formats submodule to its official repo.
This change eliminates the lag time between Bio-Formats commits and
seeing them reflected in Fiji when fetching from the submodule.

One small issue is that Bio-Formats does not have a "contrib" user set
up over SSH. Currently, the only people who can push to bio-formats.git
are those authorized to do so in the gitosis configuration. However, in
principle the BF team has no objection to enabling a contrib branch; it
is merely an issue of time and knowledge to do so. If anybody wants this
feature, let us know and we will add it!