Update to ImageJ 1.45r
Yes, this really is ImageJ, not ImageJA. ImageJA is not updated anymore
for the foreseeable future.

For this to work, we have put in a substantial amount of effort to take
out all the Fiji-specific changes from ImageJA and place them into a
run-time patching system (see the IJHacker class in Fiji.jar).

The headless support was similarly put into a run-time patching system
(see the Headless class in Fiji.jar).

This now adds javassist to the required runtime-dependencies, but that's
okay, we'll need it for ImageJ2, anyway.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>