Fiji launcher --mini-maven: fix running ImageJ2
The problem was that ij.jar was in the system class path, but the
Javassist-based runtime patching performed by ImageJ2 put new classes
into the URLClassLoader (instantiated by MiniMaven in order to find
ImageJ2 and its dependencies to begin with). But when the class
loader wants to load the ImagePlus class, it also needs to load the
CompositeImage class since it is accessed from within ImagePlus. And
the CompositeImage is actually a subclass of ImagePlus and needs to
access protected fields and methods. But since Javassist does not need
to patch CompositeImage, it loads it via the class loader mechanism,
which means the system class loader if ij.jar is on the system class
path. Now, access permissions cannot cross class loader boundaries,
so it results in a VerifyError.

The solution is simple: remove ij.jar from the system class
path. Actually, let's remove all .jar files from the system class
path except the one needed to run MiniMaven.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>