IJHacker: fix thinko in CommandListenerPlus handling
Not in all cases is commandExecuting() run before runCommand() in the
ij.Executer class. Hence we need to repeat the check whether to
instantiate a new fiji.command.Command.

While at it, make sure that the code also runs when ij.jar was compiled in
debug mode, in which case the parameter "cmd" shadows the field of the
same name.

Reported by Sebastian Rhode and John Lim, and John provided also a very
simple way to reproduce. Thanks!

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>