Miji: fix a bug in MATLAB java class path
In Miji.m, there it a call to

 fiji.User_Plugins.installScripts();

Unfortunately, this causes a nasty bug with MATLAB: calling this
static method modifies the static MATLAB java path, which is
normally forbidden. The consequences of that are nasty: adding a
class to the dynamic class path can be refused, because it would be
falsy recorded in the static path. On top of that, the static
path is fsck in a weird way, with file separator from Unix, causing a
mess on Windows platform.
So we give it up as now.
Signed-off-by: Jean-Yves Tinevez <jeanyves.tinevez@gmail.com>