TrackMate_: change the way settings are retrieved in XML
We first now load the segmenter (resp. tracker) name from the XML and
instantiate it. This segmenter gives us a default segmenter settings. If
the class name of the default settings matches the class name saved in
the XML file, then we un-marshall it. Otherwise, we keep defautl value.
This is very defensive (the saved params are retrieved only if
everything is ok) but ensures that model read from XML will always
contain data that will work if it is processed.

Signed-off-by: Jean-Yves Tinevez <jean-yves.tinevez@pasteur.fr>