Fix the old tests from VIB
These tests can be run from the root Fiji directory with:

    tests/vib.py

They depend on some sample image data, which will be downloaded
if it is missing.  There are still a couple of test failures,
but they look like genuine changes in behaviour which I'll
look at separately.

I've also switched to using JUnitCore.main directly rather
than re-invoking Fiji in a subdirectory.  We check that the
current directory is the Fiji root, and now all references
to sample data should be relative to there.