FijiClassLoader: prefer newer files from Fiji.app/update/
When the updater downloaded new files, it _might_ be safe to just take
those .jar files over the ones in plugins/ and jars/.

It is _not_ safe if changes in the Fiji launcher, ij.jar, javassist.jar
or Fiji.jar are required for the updated packages to perform properly.

So do not advertise this feature, but keep it for power users who Know
What They're Doing.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>