Fiji Updater: introduce the concept of an updateable platform
On multi-platform setups, it is convenient to have only one unpacked
fiji-nojre, but the updater still only updates the current platform.

Now we try to be clever and detect whether the user has a launcher for
a given platform and update that as well as the current one.

This is only the first part of that support, the next commit actually uses
the method isUpdateablePlatform().

Yes, "updateable" is a misspelling, but we did it all the time and nobody
complained, now we have public methods of that name.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>