TrackMate_: fix a nasty bug in TrackScheme
When showing the trackscheme with spot images, a thread was taking 100%
of the time, and bloking the refresh of the hyperstack displayer.
It took me forever to find out why.
When drawing the spot images on the TrackScheme, it seems that the
graphics method drawImage was calling itself in a loop, caused by
inheriting from some JGraphX shapes.
To fix it, I hardcoded the draw image routines in the concrete class of
mine, and this fixed the problem.

Trying to find the source for this problem, I had to review most of the
drawing part of TrackScheme, which generated some nice though little
improvment in the layout performance. TrackScheme now uses a bit less
memory and is a bit more responsive, particularly for large spot
collections.
Signed-off-by: Jean-Yves Tinevez <jean-yves.tinevez@pasteur.fr>