TrackMate_: now support multi-channel images
This is a crude update: basically, if guven a stack with multiple
channels, the segmentation and feature calculations are performed in the
currently selected channel, at the moment where the action button is
pressed.
It works, but it is crude. One may want to add some flexibility and let
the user choose the segmentation channel, and authorize to compute
features in another one.
Signed-off-by: Jean-Yves Tinevez <jeanyves.tinevez@gmail.com>