Build legacy launcher 'fiji' by default
Due to recent changes in the .jar hierarchy, the former version of the
Fiji launcher is unprepared to launch current Fiji and therefore errors
out instead of warning the user to use the ImageJ executable.

According to Postel's law, that force was misguided anyway, so let's just
copy over the ImageJ executable and not disrupt any existing workflow any
longer.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>