Add a script to copy the precompiled launchers into the correct places
This includes copying the ImageJ launchers also to 'fiji-<platform>', as
suggested by Stephan Preibisch. By doing so, we can accomodate
installations of Fiji where users created shortcuts or symbolic links
pointing to the Fiji launcher.

The plan is, however, to ship the next full Fiji release without the
fiji-<platform> files so that new installations will only have the ImageJ
launcher (the Updater will then assume that the user uninstalled the Fiji
launcher and not offer it for installation).

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>