TrackMate_: ensure the correct tracking config panel is shown
When stepping back and forth in the wizard, there might be situations
were the user would change the tracker, but the tracker settings be
kept, because it is compatible for the 2 trackers.
However a problem arise when the settings object instantiates the config
panel: we want this to match the new tracker.
This situation arises only for the LAPTracker family, handle it
specifically.
Signed-off-by: Jean-Yves Tinevez <jean-yves.tinevez@pasteur.fr>